boxing, enums, dates, big decimal

+++ ALGAWORKS/banco/src/com/algaworks/banco/modelo/pagamento/DocumentoPagavel.java
@@ -1,12 +1,14 @@
 package com.algaworks.banco.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 public interface DocumentoPagavel {
 
     // estabelecer contrato (interface)
     // automaticamente os métodos já são public and abstract
     // public abstract double getValorTotal();
 
-    double getValorTotal();
+    BigDecimal getValorTotal();
     boolean estaPago();
     void quitarPagamento();
